RC release
Added attribute emphasising ServiceCollectionFactory as not stable

Define and use LocalWebDriverFactoryBase Abstract class
