+++ src/Test/AlexanderOnTest.NetCoreWebDriverFactory.Lib.Test/AlexanderOnTest.NetCoreWebDriverFactory.Lib.Test.csproj
@@ -15,7 +15,7 @@
     <PackageReference Include="log4net" Version="2.0.8" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.2.0" />
     <PackageReference Include="Scrutor" Version="3.0.2" />
-    <PackageReference Include="Selenium.WebDriver.WebDriverFactoryNunitConfig" Version="3.0.0-beta1" />
+    <PackageReference Include="Selenium.WebDriver.WebDriverFactoryNunitConfig" Version="3.0.0-RC" />
   </ItemGroup>
 
   <ItemGroup>
